+++ src/Context.php
@@ -63,7 +63,7 @@ class Context
     {
         return property_exists($this, $type);
     }
-    
+
     /**
      * Check if a property is NOT set directly on this context and but its parent context.
      *
@@ -162,26 +162,75 @@ class Context
     }
 
     /**
+     * A short piece of text, usually one line, providing the basic function of the associated element.
      * @return string|null
      */
-    public function extractDescription()
+    public function phpdocSummary()
     {
-        $lines = explode("\n", $this->comment);
-        unset($lines[0]);
-        $description = '';
+        $content = $this->phpdocContent();
+        if (!$content) {
+            return null;
+        }
+        $lines = explode("\n", $content);
+        $summary = '';
         foreach ($lines as $line) {
+            $summary .= $line."\n";
+            if ($line === '' || substr($line, -1) === '.') {
+                return trim($summary);
+            }
+        }
+        $summary = trim($summary);
+        if ($summary === '') {
+            return null;
+        }
+        return $summary;
+    }
+    
+    /**
+     * An optional longer piece of text providing more details on the associated element’s function. This is very useful when working with a complex element.
+     * @return string|null
+     */
+    public function phpdocDescription()
+    {
+        $summary = $this->phpdocSummary();
+        if (!$summary) {
+            return null;
+        }
+        $description = trim(substr($this->phpdocContent(), strlen($summary)));
+        if ($description === '') {
+            return null;
+        }
+        return $description;
+    }
+
+    /**
+     * The text contents of the phpdoc comment (excl. tags)
+     * @return string|null
+     */
+    public function phpdocContent()
+    {
+        $comment = explode("\n", $this->comment);
+        $comment[0] = preg_replace('/[ \t]*\\/\*\*/', '', $comment[0]); // strip '/**'
+        $i = count($comment) -1;
+        $comment[$i] = preg_replace('/\*\/[ \t]*$/', '', $comment[$i]); // strip '*/'
+        $lines = [];
+        $append = false;
+        foreach ($comment as $line) {
             $line = ltrim($line, "\t *");
             if (substr($line, 0, 1) === '@') {
                 break;
             }
-            $description .= $line . ' ';
+            if ($append) {
+                $i = count($lines) - 1;
+                $lines[$i] = substr($lines[$i], 0, -1).$line;
+            } else {
+                $lines[] = $line;
+            }
+            $append = (substr($line, -1) === '\\');
         }
-        $description = trim($description);
+        $description = trim(implode("\n", $lines));
         if ($description === '') {
             return null;
-        }
-        if (stripos($description, 'license')) {
-            return null; // Don't use the GPL/MIT license text as the description.
         }
         return $description;
     }